refactor: MAT-362 PositionTag 컴포넌트 teams/players -> 공통 components로 이관
