+++ src/components/index.ts
@@ -24,3 +24,4 @@ export * from './Button';
 export * from './MatchCreateStepper';
 export * from './Table';
 export * from './PlayerReceivedCard';
+export * from './PositionTag';
