+++ src/routes/teams/$teamId/players/-components/index.ts
@@ -2,4 +2,3 @@ export * from './MemberStatsSummary';
 export * from './MemberFilters';
 export * from './MemberTable';
 export * from './PlayerDetailPanel';
-export * from './PositionTag';
